feat(cache) Document and polish the cache implementation.
