+++ lib/Cache/Exception.php
@@ -7,6 +7,9 @@ namespace Wasm\Cache;
 use Psr\SimpleCache\CacheException;
 use RuntimeException;
 
+/**
+ * Represents all errors for the cache implementations.
+ */
 class Exception extends RuntimeException implements CacheException
 {
 }
