+++ lib/Cache/InvalidArgumentException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Wasm\Cache;
+
+use Psr\SimpleCache;
+
+/**
+ * Represents an invalid argument error for the cache implementations.
+ */
+class InvalidArgumentException extends Exception
+{
+}
